Renamed NavBar component's directory.
